requirements.txt
asgiref==3.10.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
cssselect==1.3.0
cssutils==2.11.1
Django==5.2.8
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.3
google-api-python-client==2.156.0
gspread==5.12.0
idna==3.11
lxml==6.0.2
more-itertools==10.8.0
oauthlib==3.3.1
premailer==3.10.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
urllib3==2.5.0
wheel==0.45.1
yagmail==0.15.293
